Fix feedback port and power thru hole size
Feedback ethernet port was fixed by adding a missing trace.
The 24VDC in thru hole size diameter was increased to adhere to the
datasheet's specifications.
The JST plugs used for valve actuation (tanks vent and main fuel) were
change from SMD to PTH as we had problems with the SMD connectors
breaking off when connecting or disconnecting the plugs.